requirements/common.txt
# Core Stuff
# -------------------------------------
Django==1.11.1
django-flat-responsive==1.3.0

# WSGI HTTP Server
# -------------------------------------
uWSGI==2.0.15

# Configuration
# -------------------------------------
argon2-cffi==16.3.0
django-environ==0.4.3
django-sites==0.9
python-dotenv==0.6.4
django-cors-headers==2.0.2

# Staticfiles
# -------------------------------------
whitenoise==3.3.0

# Extensions
# -------------------------------------
pytz==2017.2

# Models
# -------------------------------------
psycopg2==2.7.1

Pillow==4.1.1
django-versatileimagefield==1.7.0
django-uuid-upload-path==1.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.6.3
django-rest-swagger==2.1.2

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Flickr
# --------------------------------------
flickrapi==2.3
